+++ docs/app/raw/openapi/openapi.json/route.ts
@@ -4,5 +4,13 @@ import { ir } from '@saltify/milky-protocol';
 export const dynamic = 'force-static';
 
 export async function GET() {
-  return new Response(JSON.stringify(await generateOpenApiSpec(ir, await import('@saltify/milky-types'))));
+  return new Response(
+    JSON.stringify(
+      await generateOpenApiSpec(
+        ir,
+        // @ts-expect-error
+        await import('@/zod-types.g'),
+      ),
+    ),
+  );
 }